fix(im): 修复复审第 1 轮发现的 5 处问题
复审确认前一轮的三处核心缺陷修复思路成立，但指出 5 处必须修，其中 3 处
复审者实际改代码跑测试证明了对应回归测试测不到目标缺陷。

1. Envelope.Hop 从 uint8 改成 uint16：候选数达到 256 时 uint8 会回绕成 0，
   下游把游标读成"从 Route[0] 重新开始"，导致跨节点永久打转、同一条消息
   被重复投给同一条 server 流。新增 TestDeliverRouteLargerThan255DoesNotWrapAround
   验证 300 候选场景下 Hop 值不回绕。

2. TestRemoveConnSkipsDisconnectWhenConnectedNotDelivered 抓不到删掉 C2
   第 6 步（if !entry.delivered）的情况：AddConn/RemoveConn 走的是同一条
   "候选为空"路径，两种实现下断言都成立。改成 AddConn 之后再挂本地流，
   让"建立事件确实投递失败、断开事件本应能成功但不该发"这两件事分开
   验证。

3. TestHeartbeatBackstopRestoresDroppedEntry 测不到心跳分支的
   writeServingTable 是否被调用：Beat() 本身就会 HSET serving 集合里的
   app，旧测试断言的 HSET 方向被 Beat() 单独覆盖。重写为
   TestHeartbeatBackstopHDelsStaleEntry，专测 HDEL 方向（Beat 从不 HDEL）。

4. TestSetServingOutOfOrderRegression 有约 1/8 假阳性：go l.Run(runCtx)
   的启动预热轮什么时候被调度不确定，若晚于测试的"甲乙"交错执行，会
   独立地把状态写对、掩盖被测代码其实是旧实现。新增 waitRunPrimed，
   在预热轮确认跑完（此时 everSrv 为空，预热轮内的写是空跑）之后再开始
   交错。

5. 补全 C2 契约注释：业务方处理断开事件遇到不认识的连接 id 必须安静
   忽略，因为业务 server 自身重启后也会对启动前就存在的连接收到陌生
   连接 id，这是 im 消除不掉的、必须由接入方容忍的行为。

顺带两处：writeServingTable 的 Pipelined 错误从静默吞掉改成 slog.Warn；
AddConn 的 close(entry.established) 改成 defer，避免 emit panic 时
established 永远不关闭、让并发的 RemoveConn 白等满 30 秒。

上述第 1/2/3/4 条均已回退验证：改回缺陷版本后测试变红（第 4 条用
-count=8 跑了八次全部变红），改回修复版本后重新变绿，过程见
fix-batch-1-report.md「修复轮 2」。

+++ internal/im/hub/deliver.go
@@ -66,7 +66,7 @@ func (h *Hub) Deliver(ctx context.Context, env bus.Envelope) bool {
 
 	for i := int(env.Hop); i < len(env.Route); i++ {
 		target := env.Route[i]
-		env.Hop = uint8(i + 1) // 先把游标推进到下一位再发出去，下游拿到的是"从这里继续"
+		env.Hop = uint16(i + 1) // 先把游标推进到下一位再发出去，下游拿到的是"从这里继续"
 		n, err := h.pub.Publish(ctx, target, env)
 		if err != nil {
 			slog.Warn("hub: 发布到节点失败，尝试下一候选", "target", target, "err", err)
